backend/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.2.0
